// client/src/cores/constants/enums.ts
export enum ColorType {
  primary = "#03A9F4",
  background = "#f2f2f2",
  surface = "#ffffff",
  onSurface = "#212121",
  onPrimary = "#ffffff",
  secondary = "#37474F",
  onSecondary = "#ffffff",
  error = "#F50057",
  mute = "#f9f9f9",
  onSurfaceLight = "#E0E0E0",
  primaryLight = "#29B6F6",
  secondaryLight = "#455A64",
  errorLight = "#E91E63",
  disable = "#FAFAFA",
  onDisable = "#CFD8DC",
}